+++ src/main/module/db/config-database.go
@@ -1,0 +1,11 @@
+package database
+
+import (
+	repositoryUser "go-api/src/infra/repositories/user"
+
+	"github.com/jinzhu/gorm"
+)
+
+func LoadMigrationByRepositores(db *gorm.DB) {
+	repositoryUser.InitMigrate(db)
+}